requirements.txt
openai
python-dotenv
scikit-learn
joblib
